requirements.txt
apify_client==1.8.1
Jinja2==3.1.4
openai==1.50.2
Pillow==10.4.0
praw==7.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
